requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.110.0
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.29.0
